refactor: separate concerns in peer notification and fix broadcast storage
Major refactoring of app state broadcasting:

1. Split peerNotification into broadcast-only + recovery:
   - Stopped-app recovery logic (recreateMissingContainers,
     handleMissingMasterSlaveContainer, checkStoppedApps) moved to
     stoppedAppsRecovery.js where it belongs
   - peerNotification.js now only handles broadcasting + timer
   - Breaks circular dep: appInstaller no longer imports peerNotification
   - appInstaller uses setOnInstallComplete callback wired by serviceManager

2. Always send v2 apprunning broadcasts:
   - peerNotification always sends v2 with full apps array
   - appInstaller triggers full checkAndNotifyPeersOfRunningApps via callback
   - Removes v1/v2 branching on the sending side

3. Fix v1 broadcast storage overwriting:
   - v1 broadcasts (from old nodes) now upsert by {ip, data.name} not {ip}
   - Prevents losing per-app data when multiple v1 messages arrive
   - Updated unique index to {ip, data.name}

4. Broadcast timer self-resets:
   - Moved interval from orchestrator to peerNotification
   - resetBroadcastInterval called in finally block
   - Any call to checkAndNotifyPeersOfRunningApps resets the 1h timer

5. appLocationFromBroadcasts aggregation view:
   - New function produces same output shape as appLocation()
   - Handles both v1 and v2 via $facet, dedupes by {name, ip}
   - Added data.apps.name index for v2 app queries
   - Prepares for replacing locations collection with broadcast collection
