+++ ZelBack/src/services/appMessaging/peerNotification.js
@@ -1,147 +1,44 @@
-// Peer Notification Service - Manages broadcasting of running apps to network peers
 const os = require('os');
 const config = require('config');
 const dbHelper = require('../dbHelper');
-const dockerService = require('../dockerService');
-const serviceHelper = require('../serviceHelper');
 const generalService = require('../generalService');
 const benchmarkService = require('../benchmarkService');
 const geolocationService = require('../geolocationService');
 const fluxCommunicationMessagesSender = require('../fluxCommunicationMessagesSender');
 const messageStore = require('./messageStore');
-const registryManager = require('../appDatabase/registryManager');
-const appInspector = require('../appManagement/appInspector');
-const appUninstaller = require('../appLifecycle/appUninstaller');
-const appInstaller = require('../appLifecycle/appInstaller');
 const { decryptEnterpriseApps } = require('../appQuery/appQueryService');
-const { localAppsInformation } = require('../utils/appConstants');
 const log = require('../../lib/log');
 const globalState = require('../utils/globalState');
 const appQueryService = require('../appQuery/appQueryService');
-const cacheManager = require('../utils/cacheManager').default;
-const appTamperingDetectionService = require('../appTamperingDetectionService');
+const stoppedAppsRecovery = require('../appLifecycle/stoppedAppsRecovery');
 
-// Database collections
 const globalAppsLocations = config.database.appsglobal.collections.appsLocations;
 
-// Module-level state variable
 let checkAndNotifyPeersOfRunningAppsFirstRun = true;
+let broadcastInterval = null;
 
-/**
- * Recreate containers for app that exists in DB but has missing containers
- * @param {string} componentIdentifier - Component identifier (component_appname or appname)
- * @returns {Promise<void>}
- */
-async function recreateMissingContainers(componentIdentifier) {
-  const mainAppName = componentIdentifier.split('_')[1] || componentIdentifier;
-  const dbopen = dbHelper.databaseConnection();
-  const appsDatabase = dbopen.db(config.database.appslocal.database);
-
-  const appsQuery = { name: mainAppName };
-  const appsProjection = { projection: { _id: 0 } };
-  let appSpec = await dbHelper.findOneInDatabase(appsDatabase, localAppsInformation, appsQuery, appsProjection);
-
-  if (!appSpec) {
-    throw new Error(`App ${mainAppName} not found in local database`);
-  }
-
-  appSpec = await decryptEnterpriseApps([appSpec], { formatSpecs: false });
-  appSpec = appSpec[0];
-
-  if (!appSpec.compose || appSpec.compose.length === 0) {
-    throw new Error(`App ${mainAppName} has no components to install`);
-  }
-
-  const tier = await generalService.nodeTier();
-  const isComponent = componentIdentifier.includes('_');
-
-  if (isComponent) {
-    const componentName = componentIdentifier.split('_')[0];
-    const componentSpec = appSpec.compose.find((c) => c.name === componentName);
-    if (!componentSpec) {
-      throw new Error(`Component ${componentName} not found in app ${mainAppName}`);
-    }
-    const hddTier = `hdd${tier}`;
-    const ramTier = `ram${tier}`;
-    const cpuTier = `cpu${tier}`;
-    componentSpec.cpu = componentSpec[cpuTier] || componentSpec.cpu;
-    componentSpec.ram = componentSpec[ramTier] || componentSpec.ram;
-    componentSpec.hdd = componentSpec[hddTier] || componentSpec.hdd;
-    await appInstaller.installApplicationHard(componentSpec, mainAppName, true, null, appSpec);
-  } else {
-    for (const componentSpec of appSpec.compose) {
-      const hddTier = `hdd${tier}`;
-      const ramTier = `ram${tier}`;
-      const cpuTier = `cpu${tier}`;
-      componentSpec.cpu = componentSpec[cpuTier] || componentSpec.cpu;
-      componentSpec.ram = componentSpec[ramTier] || componentSpec.ram;
-      componentSpec.hdd = componentSpec[hddTier] || componentSpec.hdd;
-      await appInstaller.installApplicationHard(componentSpec, mainAppName, true, null, appSpec);
-    }
-  }
-
-  log.info(`Successfully recreated missing containers for ${componentIdentifier}`);
+function resetBroadcastInterval() {
+  if (broadcastInterval) clearInterval(broadcastInterval);
+  broadcastInterval = setInterval(() => {
+    checkAndNotifyPeersOfRunningApps();
+  }, 60 * 60 * 1000);
 }
 
-/**
- * Handle master/slave apps with missing containers.
- * Stopped containers are normal for master/slave apps. Missing containers need recovery.
- * @param {string} stoppedApp - Component identifier
- * @param {string} mainAppName - Main app name
- * @param {object} appsMonitored - Monitored apps object
- * @param {function} getGlobalState - Global state getter
- */
-async function handleMissingMasterSlaveContainer(stoppedApp, mainAppName) {
-  const containerExists = await dockerService.getDockerContainerOnly(stoppedApp);
-  if (containerExists) return;
-
-  log.warn(`Container for master/slave app ${stoppedApp} doesn't exist, recreating...`);
-  try {
-    await recreateMissingContainers(stoppedApp);
-    log.info(`Successfully recreated master/slave app container ${stoppedApp}`);
-    appInspector.startAppMonitoring(stoppedApp, globalState.appsMonitored);
-  } catch (recreateErr) {
-    // Check if container now exists — another process (e.g. masterSlaveApps) may have created it
-    const containerExistsNow = await dockerService.getDockerContainerOnly(stoppedApp);
-    if (containerExistsNow) {
-      log.info(`Container for ${stoppedApp} was created by another process, skipping removal`);
-      return;
-    }
-    log.error(`Failed to recreate master/slave app ${stoppedApp}: ${recreateErr.message}`);
-    log.warn(`REMOVAL REASON: Master/slave container recreation failure - ${mainAppName} (peerNotification)`);
-    await appTamperingDetectionService.recordEvent(mainAppName, 'recreation_failed', `Master/slave container recreation failure: ${recreateErr.message}`);
-    if (appTamperingDetectionService.isNetworkMissingError(recreateErr.message)) {
-      await appTamperingDetectionService.recordEvent(mainAppName, 'network_pruned', `Docker network missing during recreation: ${recreateErr.message}`);
-    }
-    await appUninstaller.removeAppLocally(mainAppName, null, false, true, true, () => {},
-      () => globalState, (name, deleteData) => appInspector.stopAppMonitoring(name, deleteData, globalState.appsMonitored));
+function stopBroadcastInterval() {
+  if (broadcastInterval) {
+    clearInterval(broadcastInterval);
+    broadcastInterval = null;
   }
 }
 
-/**
- * Check and notify peers of running applications
- * This function is called periodically to broadcast the status of running apps to the network
- * @param {function} installedApps - Function to get installed apps
- * @param {function} listRunningApps - Function to get running apps
- * @param {object} appsMonitored - Object tracking monitored apps
- * @param {boolean} removalInProgress - Whether app removal is in progress
- * @param {boolean} installationInProgress - Whether app installation is in progress
- * @param {boolean} softRedeployInProgress - Whether soft redeploy is in progress
- * @param {boolean} hardRedeployInProgress - Whether hard redeploy is in progress
- * @param {boolean} reinstallationOfOldAppsInProgress - Whether reinstallation is in progress
- * @param {function} getGlobalState - Function to get global state
- * @param {object} cacheManager - Cache manager instance with stoppedAppsCache
- */
 async function checkAndNotifyPeersOfRunningApps() {
   try {
-    let isNodeConfirmed = false;
-    isNodeConfirmed = await generalService.isNodeStatusConfirmed().catch(() => null);
+    const isNodeConfirmed = await generalService.isNodeStatusConfirmed().catch(() => null);
     if (!isNodeConfirmed) {
       log.info('checkAndNotifyPeersOfRunningApps - FluxNode is not Confirmed');
       return;
     }
 
-    // get my external IP and check that it is longer than 5 in length.
     const benchmarkResponse = await benchmarkService.getBenchmarks();
     let myIP = null;
     if (benchmarkResponse.status === 'success') {
@@ -154,8 +51,7 @@ async function checkAndNotifyPeersOfRunningApps() {
     if (myIP === null) {
       throw new Error('Unable to detect Flux IP address');
     }
-    // get list of locally installed apps. Store them in database as running and send info to our peers.
-    // check if they are running?
+
     const installedAppsRes = await appQueryService.installedApps();
     if (installedAppsRes.status !== 'success') {
       throw new Error('Failed to get installed Apps');
@@ -167,134 +63,15 @@ async function checkAndNotifyPeersOfRunningApps() {
     let appsInstalled = installedAppsRes.data;
     appsInstalled = await decryptEnterpriseApps(appsInstalled, { formatSpecs: false });
     const runningApps = runningAppsRes.data;
-    const installedAppComponentNames = [];
-    appsInstalled.forEach((app) => {
-      if (app.version >= 4) {
-        app.compose.forEach((appComponent) => {
-          installedAppComponentNames.push(`${appComponent.name}_${app.name}`);
-        });
-      } else {
-        installedAppComponentNames.push(app.name);
-      }
-    });
-    // kadena and folding is old naming scheme having /zel.  all global application start with /flux
     const runningAppsNames = runningApps.map((app) => {
       if (app.Names[0].startsWith('/zel')) {
         return app.Names[0].slice(4);
       }
       return app.Names[0].slice(5);
     });
-    // installed always is bigger array than running
-    const runningSet = new Set(runningAppsNames);
-    const stoppedApps = installedAppComponentNames.filter((installedApp) => !runningSet.has(installedApp));
-    const masterSlaveAppsInstalled = [];
 
-    const backupInProgress = globalState.backupInProgress || [];
-    const restoreInProgress = globalState.restoreInProgress || [];
-    const appsStopedCache = cacheManager.stoppedAppsCache;
+    const masterSlaveAppsInstalled = await stoppedAppsRecovery.checkStoppedApps(myIP, appsInstalled, runningAppsNames);
 
-    // check if stoppedApp is a global application present in specifics. If so, try to start it.
-    if (!globalState.removalInProgress && !globalState.installationInProgress && !globalState.softRedeployInProgress && !globalState.hardRedeployInProgress && !globalState.reinstallationOfOldAppsInProgress) {
-      // eslint-disable-next-line no-restricted-syntax
-      for (const stoppedApp of stoppedApps) { // will uninstall app if some component is missing
-        try {
-          // proceed ONLY if it's a global App
-          const mainAppName = stoppedApp.split('_')[1] || stoppedApp;
-          // eslint-disable-next-line no-await-in-loop
-          const appDetails = await registryManager.getApplicationGlobalSpecifications(mainAppName);
-          const appInstalledMasterSlave = appsInstalled.find((app) => app.name === mainAppName);
-          const appInstalledSyncthing = appInstalledMasterSlave.compose.find((comp) => comp.containerData.includes('g:') || comp.containerData.includes('r:'));
-          const appInstalledMasterSlaveCheck = appInstalledMasterSlave.compose.find((comp) => comp.containerData.includes('g:'));
-          if (appInstalledSyncthing) {
-            masterSlaveAppsInstalled.push(appInstalledMasterSlave);
-          }
-          if (appInstalledMasterSlaveCheck && appDetails) {
-            const backupSkip = backupInProgress.some((backupItem) => stoppedApp === backupItem);
-            const restoreSkip = restoreInProgress.some((backupItem) => stoppedApp === backupItem);
-            if (!backupSkip && !restoreSkip) {
-              // eslint-disable-next-line no-await-in-loop
-              await handleMissingMasterSlaveContainer(stoppedApp, mainAppName);
-            }
-          } else if (appDetails) {
-            log.warn(`${stoppedApp} is stopped but should be running. Starting...`);
-            // it is a stopped global app. Try to run it.
-            // check if some removal is in progress and if it is don't start it!
-            const backupSkip = backupInProgress.some((backupItem) => stoppedApp === backupItem);
-            const restoreSkip = restoreInProgress.some((backupItem) => stoppedApp === backupItem);
-            if (backupSkip || restoreSkip) {
-              log.warn(`Application ${stoppedApp} backup/restore is in progress...`);
-            }
-            if (!globalState.removalInProgress && !globalState.installationInProgress && !globalState.softRedeployInProgress && !globalState.hardRedeployInProgress && !globalState.reinstallationOfOldAppsInProgress && !restoreSkip && !backupSkip) {
-              // eslint-disable-next-line no-await-in-loop
-              const containerExists = await dockerService.getDockerContainerOnly(stoppedApp);
-
-              // Check if container exists before applying syncthing delay
-              if (containerExists && appInstalledSyncthing) {
-                const db = dbHelper.databaseConnection();
-                const database = db.db(config.database.appsglobal.database);
-                const queryFind = { name: mainAppName, ip: myIP };
-                const projection = { _id: 0, runningSince: 1 };
-                // we already have the exact same data
-                // eslint-disable-next-line no-await-in-loop
-                const result = await dbHelper.findOneInDatabase(database, globalAppsLocations, queryFind, projection);
-                if (!result || !result.runningSince || Date.parse(result.runningSince) + 30 * 60 * 1000 > Date.now()) {
-                  log.info(`Application ${stoppedApp} uses r syncthing and container exists but is stopped. Haven't started yet because was installed less than 30m ago.`);
-                  // eslint-disable-next-line no-continue
-                  continue;
-                }
-              }
-
-              if (!containerExists) {
-                log.warn(`Container for ${stoppedApp} doesn't exist, recreating immediately...`);
-                // eslint-disable-next-line no-await-in-loop
-                await appTamperingDetectionService.recordEvent(mainAppName, 'container_vanished', `Container ${stoppedApp} missing, not found in Docker`);
-                try {
-                  // eslint-disable-next-line no-await-in-loop
-                  await recreateMissingContainers(stoppedApp);
-                  log.info(`Successfully recreated ${stoppedApp}`);
-                  appInspector.startAppMonitoring(stoppedApp, globalState.appsMonitored);
-                } catch (recreateErr) {
-                  log.error(`Failed to recreate containers for ${stoppedApp}: ${recreateErr.message}`);
-                  log.warn(`REMOVAL REASON: Container recreation failure - ${mainAppName} failed to recreate with error: ${recreateErr.message} (peerNotification)`);
-                  // eslint-disable-next-line no-await-in-loop
-                  await appTamperingDetectionService.recordEvent(mainAppName, 'recreation_failed', `Container recreation failure: ${recreateErr.message}`);
-                  if (appTamperingDetectionService.isNetworkMissingError(recreateErr.message)) {
-                    // eslint-disable-next-line no-await-in-loop
-                    await appTamperingDetectionService.recordEvent(mainAppName, 'network_pruned', `Docker network missing during recreation: ${recreateErr.message}`);
-                  }
-                  // eslint-disable-next-line no-await-in-loop
-                  await appUninstaller.removeAppLocally(mainAppName, null, false, true, true, () => {
-                    // Handle response
-                  }, () => globalState, (name, deleteData) => appInspector.stopAppMonitoring(name, deleteData, globalState.appsMonitored));
-                }
-              } else {
-                log.warn(`${stoppedApp} is stopped, starting`);
-                if (!appsStopedCache.has(stoppedApp)) {
-                  appsStopedCache.set(stoppedApp, '');
-                } else {
-                  // eslint-disable-next-line no-await-in-loop
-                  await dockerService.appDockerStart(stoppedApp);
-                  appInspector.startAppMonitoring(stoppedApp, globalState.appsMonitored);
-                }
-              }
-            } else {
-              log.warn(`Not starting ${stoppedApp} as application removal or installation or backup/restore is in progress`);
-            }
-          }
-        } catch (err) {
-          log.error(err);
-          if (!globalState.removalInProgress && !globalState.installationInProgress && !globalState.softRedeployInProgress && !globalState.hardRedeployInProgress && !globalState.reinstallationOfOldAppsInProgress) {
-            log.warn(`REMOVAL REASON: App start failure - ${mainAppName} failed to start with error: ${err.message} (peerNotification)`);
-            // eslint-disable-next-line no-await-in-loop
-            await appUninstaller.removeAppLocally(mainAppName, null, false, true, true, () => {
-              // Handle response
-            }, () => globalState, (name, deleteData) => appInspector.stopAppMonitoring(name, deleteData, globalState.appsMonitored));
-          }
-        }
-      }
-    } else {
-      log.warn('Stopped application checks not running, some removal or installation is in progress');
-    }
     const installedAndRunning = [];
     appsInstalled.forEach((app) => {
       if (app.version >= 4) {
@@ -321,9 +98,9 @@ async function checkAndNotifyPeersOfRunningApps() {
       for (const application of applicationsToBroadcast) {
         const queryFind = { name: application.name, ip: myIP };
         const projection = { _id: 0, runningSince: 1 };
-        let runningOnMyNodeSince = new Date().toISOString();
         // eslint-disable-next-line no-await-in-loop
         const result = await dbHelper.findOneInDatabase(database, globalAppsLocations, queryFind, projection);
+        let runningOnMyNodeSince = new Date().toISOString();
         if (result && result.runningSince) {
           runningOnMyNodeSince = result.runningSince;
         }
@@ -352,9 +129,7 @@ async function checkAndNotifyPeersOfRunningApps() {
       log.info(`App Running Message broadcasted: ${apps.length} apps`);
     } catch (err) {
       log.error(err);
-      // removeAppLocally(stoppedApp);
     }
-    // Update the running apps cache with the current state
     const runningAppsCache = globalState.runningAppsCache;
     runningAppsCache.clear();
     apps.forEach((app) => {
@@ -364,10 +139,12 @@ async function checkAndNotifyPeersOfRunningApps() {
     log.info('Running Apps broadcasted');
   } catch (error) {
     log.error(error);
+  } finally {
+    resetBroadcastInterval();
   }
 }
 
 module.exports = {
   checkAndNotifyPeersOfRunningApps,
-  handleMissingMasterSlaveContainer,
+  stopBroadcastInterval,
 };
